Chore: Moved models to models.rs
